buildfix

+++ WowPacketParserModule.V8_0_1_27101/Parsers/CharacterHandler.cs
@@ -106,7 +106,7 @@ namespace WowPacketParserModule.V8_0_1_27101.Parsers
 
             if (firstLogin)
             {
-                PlayerCreateInfo startPos = new PlayerCreateInfo { Race = race, Class = klass, Map = mapId, Zone = zone, Position = pos, Orientation = 0 };
+                PlayerCreateInfo startPos = new PlayerCreateInfo { Race = race, Class = klass, Map = (uint)mapId, Zone = (uint)zone, Position = pos, Orientation = 0 };
                 Storage.StartPositions.Add(startPos, packet.TimeSpan);
             }
         }
